feat: settings section to change the indexer URL + new icon (#18)
Up until now the Indexer URL parameter was static, set up in the
`TARI_INDEXER_URL` envvar. This means that after building, the snap was
tied up to that indexer without a way to change it by the user during
execution. With this PR we add a new section (under Metamask settings ->
Snap -> Tari Snap) that allows the user to change the URL of the indexer
to be used. The value is persisted in the metamask snap storage. We
still have a envvar (`DEFAULT_TARI_INDEXER_URL`) for the initial indexer
URL value on a fresh new install.

This PR also updates the snap icon to use the Tari logo.

<img width="361" alt="Screenshot 2024-05-23 at 10 44 37"
src="https://github.com/tari-project/tari-snap/assets/47919901/998d3704-b0da-4262-a124-d35452b7eac1">